+++ CHANGELOG.md
@@ -7,15 +7,21 @@
 - Removed Graph Defitions
 - Removed Depreacated Structs
 - schema and type for train_length to be consistent with the database schema
+- `locations::graph` yeetus-deletus
+- Region moved to `structs`
+- InsertRegion moved to `structs`
+- type renamed `R09Types` -> `R09Type`
+- `R09Type` moved to `telegrams::R09`
+- `R09SaveTelegram.telegram_type` now uses `R09Type` instead of `i64`
 
 ### Added
 
 - Waypoints
 - Chemo Service
-
-#### DB Tables
+- custom diesel \[de\]serializer for the `R09Type`
 - r09_transmission_locations
 - r09_transmission_locations_raw
+- `TransmissionLocation`, `TransmissionLocationRaw`, `InsertTransmissionLocationRaw`, `InsertTransmissionLocation` for tables above
 
 ### Fixed